app/auth/oauth2.py
from jose import JWTError, jwt
from datetime import datetime, timedelta
from typing import Optional

# * SECRET KEY
# * ALGORITHM
# * EXPIRATION TIME

SECRET_KEY = "09d25e094faa6ca2556c818166b7a9563b93f7099f6f0f4caa6cf63b88e8d3e7"
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30


def create_access_token(
    data: dict,
    expires_delta: Optional[timedelta] = None,
):
    """Create a session access token

    Args:
        data (dict): token payload
        expires_delta (Optional[timedelta], optional): Token expiration time in minutes. Defaults to None.

    Returns:
        str: encoded JWT token
    """
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.now() + expires_delta
    else:
        expire = datetime.now() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})

    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt
